+++ src/SkynetServer.Shared/Database/Entities/MessageDependency.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace SkynetServer.Entities
+namespace SkynetServer.Database.Entities
 {
     public class MessageDependency
     {
