+++ src/SkynetServer.Web/Startup.cs
@@ -8,7 +8,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using SkynetServer.Entities;
+using SkynetServer.Database;
 
 namespace SkynetServer.Web
 {
